@@ -68,16 +68,16 @@
     "node-sass": "^3.2.0",
     "npm-run-all": "^1.2.6",
     "postcss-cli": "^1.5.0",
-    "react": "^15.0",
-    "react-addons-test-utils": "^15.0",
-    "react-dom": "^15.0",
+    "react": "^15.5",
+    "react-addons-test-utils": "^15.5",
+    "react-dom": "^15.5",
     "rimraf": "^2.3.4",
     "sinon": "^1.15.4",
     "sinon-chai": "^2.8.0",
     "webpack": "^1.11.0"
   },
   "peerDependencies": {
-    "react": "^0.14 || ^15.0",
-    "react-dom": "^0.14 || ^15.0"
+    "react": "^0.14 || ^15.5",
+    "react-dom": "^0.14 || ^15.5"
   }
 }
